+++ packages/eyes-cypress/test/fixtures/testApp/cypress/integration-run/layout-breakpoints.js
@@ -49,4 +49,20 @@ describe('JS layout', () => {
     cy.eyesCheckWindow('layoutBreakpoints = true');
     cy.eyesClose();
   });
+
+  it('should not hit the cypress default command timeout', () => {
+    cy.visit('http://localhost:5555/breakpoints.html');
+    cy.eyesOpen({
+      appName: 'JS layout',
+      testName: 'testing js layout support in cypress',
+      browser: [
+        {name: 'chrome', width: 1000, height: 800},
+        {iosDeviceInfo: {deviceName: 'iPad (7th generation)'}},
+        {chromeEmulationInfo: {deviceName: 'Pixel 4 XL'}},
+      ],
+      layoutBreakpoints: true,
+    });
+    cy.eyesCheckWindow('cypress default command timeout');
+    cy.eyesClose();
+  });
 });