+++ packages/eyes-cypress/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- set explicit timeout for eyesCheckWindow [trello](https://trello.com/c/5Thz1QFs)
 
 ## 3.20.3 - 2021/3/15